+++ CHANGELOG.md
@@ -5,10 +5,9 @@ This file is used to list changes made in each version of the zabbix cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of zabbix
+- Richardson Lima - Initial release of zabbix
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.
 
 The [Github Flavored Markdown page](http://github.github.com/github-flavored-markdown/) describes the differences between markdown on github and standard markdown.
-